EndpointProvider and fixing auth
