+++ HobbyMatchApp/Components/Pages/SignUp.razor
@@ -2,8 +2,9 @@
 
 @using System.Text.RegularExpressions
 @using System.ComponentModel.DataAnnotations
+@using HobbyMatch.App.Services.Api
 
-@inject Services.IAuthApiService AuthApiService
+@inject IAuthApiService AuthApiService
 @inject NavigationManager navigationManager
 
 <MudContainer Style="justify-content: center" MaxWidth="MaxWidth.Small">
