+++ HobbyMatchApp/Services/EndpointProvider.cs
@@ -1,0 +1,21 @@
+﻿using HobbyMatch.Domain.Entities;
+
+namespace HobbyMatch.App.Services
+{
+	public class EndpointProvider
+	{
+		private readonly Dictionary<Type, string> _endpointMap = new() 
+		{
+			{ typeof(User), "/users" },
+			{ typeof(BusinessClient), "/business-clients" },
+		};
+		public string GetEndpoint<T>()  where T : Organizer
+		{
+			if (_endpointMap.TryGetValue(typeof(T), out var endpoint))
+			{
+				return endpoint;
+			}
+			throw new InvalidOperationException($"No endpoint declared for type: {typeof(T).Name}");
+		}
+	}
+}
